FEAT divide elections info into tabs
Acceso elecciones: shows the main info and the access to each elections
Resumen elecciones: shows the resume of each election (has more info than aecceso elecciones)
